requirements.txt
moviepy==2.1.2
pyannote.audio==3.3.2
google-generativeai==0.8.4
google-genai==1.7.0
demucs==4.0.1
argparse==1.4.0
f5-tts==1.1.0